supply-chain-tracker/requirements.txt
# Core dependencies
pandas==2.0.3
numpy==1.24.4
sqlalchemy==2.0.23
pydantic==2.4.2
python-dotenv==1.0.0

# Data visualization
matplotlib==3.7.3
seaborn==0.12.2
plotly==5.18.0
dash==2.13.0

# Machine learning for forecasting
scikit-learn==1.3.2
prophet==1.1.4
tensorflow==2.13.0

# API and web framework
fastapi==0.104.1
uvicorn==0.23.2
requests==2.31.0

# Database connectors
psycopg2-binary==2.9.9
pymongo==4.6.0

# Testing
pytest==7.4.3
pytest-cov==4.1.0

# Utilities
tqdm==4.66.1
networkx==3.2.1  # For supply chain network analysis
geopy==2.4.0     # For geographical calculations 